+++ library/Vanilla/Models/SearchRecordTypeProvider.php
@@ -5,14 +5,14 @@
  * @license GPL-2.0-only
  */
 
-namespace Vanilla\Models\SearchRecords;
+namespace Vanilla\Models;
 
 use Vanilla\Contracts\Search\SearchRecordTypeInterface;
 use Vanilla\Contracts\Search\SearchRecordTypeProviderInterface;
 
 /**
  * Class SearchRecordTypeProvider
- * @package Vanilla\AdvancedSearch\Models
+ * @package Vanilla\Models
  */
 class SearchRecordTypeProvider implements SearchRecordTypeProviderInterface {
     /** @var $searchRecordTypes SearchRecordTypeInterface[] */
